おそらく動くところまで書き換えたヤツ
